+++ website/src/components/board-footer/board-footer.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import { repositoryUrl } from '../../site';
+	import { boardStyle, boardStyles } from '../../state/board-style.svelte';
 </script>
 
 <div class="board-footer">
@@ -10,15 +11,14 @@
 	</div>
 	<div class="footer-meta">
 		<span>All times are UTC</span>
-		<span class="board-style">
+		<label class="board-style">
 			Board style:
-			<span class="style-chip">
-				Vulkan Classic
-				<svg width="7" height="5" viewBox="0 0 8 5" aria-hidden="true">
-					<path d="M0 0 L4 5 L8 0 z" />
-				</svg>
-			</span>
-		</span>
+			<select bind:value={() => boardStyle.current, (id) => boardStyle.select(id)}>
+				{#each boardStyles as style (style.id)}
+					<option value={style.id}>{style.label}</option>
+				{/each}
+			</select>
+		</label>
 	</div>
 </div>